+++ src/modules/wordle.js
@@ -95,6 +95,7 @@ guess = async (guess, interaction, playNewBtn) => {
 }
 
 stats = async (user, interaction) => {
+  if(user == null){ user = interaction.user; }
   if(user.bot) return interaction.reply({embeds: [embeds.error("You can not get the stats of a bot!")], ephemeral: true});
   const data = await db.getStats(user.id).catch(_ => { return null; });
   if(!data) return interaction.reply({embeds: [embeds.error("This user unfortunatly doesn't have any stats yet!")], ephemeral: true});